% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_regression.R
\name{Debiased_inference}
\alias{Debiased_inference}
\title{Title}
\usage{
Debiased_inference(
  Y,
  A,
  W,
  mu,
  g,
  bw.seq = NULL,
  eval.pts = NULL,
  tau = 1,
  kernel.type = "epa"
)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{kernel.type}{}
}
\value{

}
\description{
Title
}
