% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_functions.R
\name{.compute.rinfl.func.quad}
\alias{.compute.rinfl.func.quad}
\title{Title: .compute.rinfl.func.quad}
\usage{
.compute.rinfl.func.quad(Y, A, a, h, b, kern, mu.hat, theta.hat)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of estimated pseudo-outcomes.}

\item{A}{\code{n x 1} numeric vector of observed exposure values.}

\item{a}{a scalar corresponding to a point of exposure to run inference.}

\item{h}{the bandwidth parameter for the local linear estimator.}

\item{b}{the bandwidth parameter for the bias estimation.}

\item{kern}{function R to R corresponding to the kernel}

\item{mu.hat}{estimated mu(A, W) as matrix. output from .pseudo.outcomes in
helper.R.}

\item{theta.hat}{estimated int mu(A, W) dW. output from .pseudo.outcomes in
helper.R.}
}
\value{

}
\description{
Title: .compute.rinfl.func.quad
}
