% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisances.R
\name{.fit.density}
\alias{.fit.density}
\title{Title: .fit.density}
\usage{
.fit.density(
  A,
  W,
  SL.library = c("SL.mean", "SL.earth", "SL.gam", "SL.rpart", "SL.glm"),
  cdf = TRUE
)
}
\arguments{
\item{A}{\code{n x 1} numeric vector of observed exposure values.}

\item{W}{\code{n x p} matrix of potential confounders.}

\item{SL.library}{list of base learners for SuperLearner package}

\item{cdf}{TRUE if it transforms the exposure A to \code{[0,1]} by empirical CDF
the default is TRUE.}
}
\value{

}
\description{
Estimate standardized propensity score using SuperLearner-based model.
The implementation is by Ted and the original method was proposed in the
paper by Diaz and van der Laan
}
