% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nuisances.R
\name{.fit.regression}
\alias{.fit.regression}
\title{Title: .fit.regression
Run outcome regression based on SuperLearner package and returns new
function that generates values for new points}
\usage{
.fit.regression(
  Y,
  A,
  W,
  SL.library = c("SL.earth", "SL.gam", "SL.glm", "SL.glmnet", "SL.glm.interaction",
    "SL.mean"),
  cdf = TRUE
)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{A}{\code{n x 1} numeric vector of observed exposure values.}

\item{W}{\code{n x p} matrix of potential confounders.}

\item{SL.library}{list of base learners for SuperLearner package}

\item{cdf}{TRUE if it transforms the exposure A to\code{[0,1]} by empirical CDF
the default is TRUE.}
}
\value{

}
\description{
Title: .fit.regression
Run outcome regression based on SuperLearner package and returns new
function that generates values for new points
}
\examples{
# Sample data
n <- 200; cols <- 3
W <- matrix(runif(n*cols), ncol=cols) # 200 * 3 matrix of covariates
A <- rnorm(n, mean=W\%*\%rnorm(cols))   # 200 * 1 vector of treatment variable
Y <- rnorm(n, mean=sin(A)+W\%*\%rnorm(cols))  # 200 * 1 vector of response
mu.hat <- .fit.regression(Y, A, W)
mu.hat(A, W) # estimated values
}
