% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_polynomials.R
\name{.lprobust.quad}
\alias{.lprobust.quad}
\title{Title: .lprobust.quad}
\usage{
.lprobust.quad(x, y, h, b, eval.pt = NULL, kernel.type = "epa")
}
\arguments{
\item{x}{\code{n x 1} numeric vector of observed covariate values.}

\item{y}{\code{n x 1} numeric vector of observed outcome values.}

\item{h}{the bandwidth parameter for the local linear estimator}

\item{b}{the bandwidth parameter for the bias estimation}

\item{eval.pt}{the sequence of points on which the regression values are
estimated}

\item{kernel.type}{The choice of a kernel function. It must be one
of the following: \code{epa} (epanechnikov), \code{tri} (triangular),
\code{uni} (uniform) and \code{gau} (Gaussian). The default is \code{epa}.}
}
\value{

}
\description{
Title: .lprobust.quad
}
