% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_regression.R
\name{debiased_ate_inference}
\alias{debiased_ate_inference}
\title{Compute debiased nonparametric inference for the difference of
covariate-adjusted regression function values.}
\usage{
debiased_ate_inference(
  Y,
  A,
  W,
  tau = 1,
  eval.pts.1 = NULL,
  eval.pts.2 = NULL,
  ...
)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{A}{\code{n x 1} numeric vector of observed exposure values.}

\item{W}{\code{n x p} data.frame of potential confounders.}

\item{tau}{the ratio of two parameters h/b}

\item{eval.pts.1}{the list of points to run inference on theta(a1)-theta(a2),
where this entry corresponds to a1. the default is 30 equi-distant points
from the 50th and 95th percentile of the exposure values}

\item{eval.pts.2}{the list of points to run inference on theta(a1)-theta(a2),
where this entry corresponds to a2. the default is the 5th percentile
of the exposure values}

\item{...}{Additional control parameters. See the \strong{Details} section.}
}
\value{

}
\description{
This function performs nonparametric inference on the difference of
causal dose-response curve values associated with continuous exposure.
The resulting confidence intervals attain asymptotically valid coverage
without undersmoothing subject to technical conditions.
See the accompanying paper for details.
}
\details{
The following is a list of additional control parameters.

\describe{
\item{\code{mu}}{outcome regression function (takes A and W and returns R)}
\item{\code{g}}{standardized propensity function
(takes A and W and returns R)}.
\item{\code{kernel.type}}{The choice of a kernel function. It must be one
of the following: \code{epa} (epanechnikov), \code{tri} (triangular),
\code{uni} (uniform) and \code{gau} (Gaussian). The default is \code{epa}.}
\item{\code{bandwidth.method}}{The choice of bandwidth selection method.
Is must be one of the following: \code{DPI} (Direct Plug-in),
\code{LOOCV} (Leave-one-out cross-validation), and
\code{LOOCV(h=b) (Leave-one-out cross-validation with the constraint h=b)}.
The default is \code{DPI}.}
\item{\code{alpha.pts}}{The alpha value for the point-wise confidence
intervals. The default value is 0.05.}
\item{\code{bw.seq}}{The set of bandwidth parameters to choose from. Only
used when \code{bandwidth.method} is \code{LOOCV} or \code{LOOCV(h=b)}.
The default value is \code{seq(0.1*n^{-1/5}, 1*n^{-1/5}, length.out = 10)} for
\code{LOOCV} and \code{seq(0.1*n^{-1/5}, 1*n^{-1/5}, length.out = 50)} for
\code{LOOCV(h=b)}}
}
}
\examples{
# Sample data
n <- 200; cols <- 3
W <- matrix(runif(n*cols), ncol=cols) # 200 * 3 matrix of covariates
A <- rnorm(n, mean=W\%*\%rnorm(cols))   # 200 * 1 vector of treatment variable
Y <- rnorm(n, mean = A^2 + rnorm(n))  # 200 * 1 vector of response variable
res <- debiased_ate_inference(Y, A, W)
}
