% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debiased_regression.R
\name{debiased_inference}
\alias{debiased_inference}
\title{Compute debiased nonparametric inference for a covariate-adjusted regression
function.}
\usage{
debiased_inference(Y, A, W, tau = 1, eval.pts = NULL, ...)
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{A}{\code{n x 1} numeric vector of observed exposure values.}

\item{W}{\code{n x p} data.frame of potential confounders.}

\item{tau}{the ratio of two parameters h/b}

\item{eval.pts}{the list of points to run inference on. the default is 30
equi-distant points from the 5th and 95th percentile of the exposure values}

\item{...}{Additional control parameters. See the \strong{Details} section.}
}
\value{

}
\description{
This function performs nonparametric inference on the causal dose-response
curve associated with continuous exposure. It performs both
pointwise and uniform inference. The resulting confidence intervals and
confidence bands attain asymptotically valid coverage without undersmoothing
subject to technical conditions.
See the accompanying paper for details.
}
\details{
The following is a list of additional control parameters.

\describe{
\item{\code{mu}}{outcome regression function (takes A and W and returns R)}
\item{\code{g}}{standardized propensity function (takes A and W and returns R)}
\item{\code{kernel.type}}{The choice of a kernel function. It must be one
of the following: \code{epa} (epanechnikov), \code{tri} (triangular),
\code{uni} (uniform) and \code{gau} (Gaussian). The default is \code{epa}.}
\item{\code{bandwidth.method}}{The choice of bandwidth selection method.
Is must be one of the following: \code{DPI} (Direct Plug-in),
\code{LOOCV} (Leave-one-out cross-validation), and
\code{LOOCV(h=b) (Leave-one-out cross-validation with the constraint h=b)}.
The default is \code{DPI}.}
\item{\code{alpha.pts}}{The alpha value for the point-wise confidence
intervals. The default value is 0.05.}
\item{\code{unif}}{If TRUE, the function performs uniform inference.
The default value is TRUE.}
\item{\code{alpha.unif}}{The alpha value for the uniform confidence bands.
The default value is 0.05.}
\item{\code{bw.seq}}{The set of bandwidth parameters to choose from. Only
used when \code{bandwidth.method} is \code{LOOCV} or \code{LOOCV(h=b)}.
The default value is \code{seq(0.1*n^{-1/5}, 1*n^{-1/5}, length.out = 10)} for
\code{LOOCV} and \code{seq(0.1*n^{-1/5}, 1*n^{-1/5}, length.out = 50)} for
\code{LOOCV(h=b)}}
\item{\code{bootstrap}}{The number of bootstrap samples used to simulate
Gaussian Process in the uniform bands. The default value is 1e4.}
}
}
\examples{
# Sample data
n <- 200; cols <- 3
W <- matrix(runif(n*cols), ncol=cols) # 200 * 3 matrix of covariates
A <- rnorm(n, mean=W\%*\%rnorm(cols))   # 200 * 1 vector of treatment variable
Y <- rnorm(n, mean = A^2 + rnorm(n))  # 200 * 1 vector of response variable
res <- debiased_inference(Y, A, W)
}
